requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
